==> 1_data-preparation.ipynb <==
### Importing the library that will be used in this Data Preparation

#importing the library that will be used
import pandas as pd
import numpy as np
import joblib
from sklearn.feature_selection import VarianceThreshold
from sklearn.model_selection import train_test_split

IMPORT THE DATASET

dataset = pd.read_csv("GSE4115_data_set.csv")

## Cleaning the data (Remove suspect sample, endcode the target variable, etc)

#rename Unnamed colomn from 'Unnamed: 0' into 'sample number'
dataset.rename(columns= {'Unnamed: 0' : 'sample_number'}, inplace=True) 

# remove sample words
dataset["Type"] = dataset["Type"].str.replace(r'Sample', '')

# remove sample number
dataset["Type"] = dataset["Type"].str.replace(r'[0-9]', '')

# remove space between target variable
dataset["Type"] = dataset["Type"].str.replace(r' ', '')

#remove suspect sample
dataset = dataset.drop(dataset.index[dataset["Type"] == 'Smokerwithsuspectlungcancer'])

#Encode The target variable
dataset['Type'] = dataset['Type'].astype('category')
dataset['Type'] = dataset['Type'].cat.codes

dataset.head(5)

## split to train and test with value 0.5
train, test = train_test_split(dataset, test_size=0.5, random_state=42)
train.shape, test.shape

train.to_csv(r'dataset-after-preparation\train.csv', index=False)
test.to_csv(r'dataset-after-preparation\test.csv', index=False)

X = dataset.drop("Type",1)
y = dataset["Type"]

X.head(5)

y.head(5)

# Remove the sample number and the index, then show the data
X = X.iloc[:,2:]
X.head(5)

#Seeing the target variable
y.head(5)

### Split into feature and target variable

# Split dataset into X_train, X_test, y_train, y_test with train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)
X_train.shape, X_test.shape, y_train.shape, y_test.shape

#Export the dataset
X_train.to_csv(r'dataset-after-preparation\X_train.csv', index=False)
y_train.to_csv(r'dataset-after-preparation\y_train.csv', index=False)
X_test.to_csv(r'dataset-after-preparation\X_test.csv', index=False)
y_test.to_csv(r'dataset-after-preparation\y_test.csv', index=False)
