==> 2_data-preprocessing.ipynb <==
Data Preprocessing (Handle duplicate, nan value, etc)

import pandas as pd
import numpy as np

#import the dataset
train_dataset = pd.read_csv('dataset-after-preparation/train_dataset.csv')
test_dataset = pd.read_csv('dataset-after-preparation/test_dataset.csv')

#Count NaN values under an entire DataFrame
train_dataset.isna().sum().sum()

#find duplicated data in dataframe 
print(train_dataset[train_dataset.duplicated()])

#find the standard deviation that more that 0.5
StdDev = ((np.std(train_dataset) > 0.5) == True)
print(StdDev)
