==> 4_build-model-SVM.ipynb <==
### Build model by using Support Vector Machine with RBF, Polynomial, and linear kernel

#### Support Vector Machine

# import required libraries
import numpy as np
import pandas as pd
# import SVC classifier
from sklearn.svm import SVC
# import metrics to compute accuracy
from sklearn.metrics import accuracy_score

#import the dataset
X_train = pd.read_csv('dataset-after-preparation/X_train.csv')
X_test = pd.read_csv('dataset-after-preparation/X_test.csv')
y_train = pd.read_csv('dataset-after-preparation/y_train.csv')
y_test = pd.read_csv('dataset-after-preparation/y_test.csv')

y_test = y_test.iloc[:,-1]
X_test = X_test.iloc[:,2:]

y_train = y_train.iloc[:,-1]
X_train = X_train.iloc[:,2:]

## SVM with default RBF Kernel 

    # instantiate classifier with rbf kernel and C=100
    svc=SVC(C=1000.0) 
    # fit classifier to training set
    svc.fit(X_train,y_train)
    # make predictions on test set
    y_pred=svc.predict(X_test)
    # compute and print accuracy score
    print('Model accuracy score with rbf kernel and C=1000.0 : {0:0.4f}'. format(accuracy_score(y_test, y_pred)))   

### SVM Using Linear Kernel 

   # instantiate classifier with linear kernel and C=1.0
   linear_svc=SVC(kernel='linear', C=1.0) 
   # fit classifier to training set
   linear_svc.fit(X_train,y_train)
   # make predictions on test set
   y_pred_test=linear_svc.predict(X_test)
   # compute and print accuracy score
   print('Model accuracy score with linear kernel and C=1.0 : {0:0.4f}'. format(accuracy_score(y_test, y_pred))) 

### SVM Using Polynomial Kernel 

    poly_svc=SVC(kernel='poly', C=1.0) 
    # fit classifier to training set
    poly_svc.fit(X_train,y_train)
    # make predictions on test set
    y_pred=poly_svc.predict(X_test)
    # compute and print accuracy score
    print('Model accuracy score with poly kernel and C=1.0 : {0:0.4f}'. format(accuracy_score(y_test, y_pred)))
