==> 6_model-validation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import joblib
from sklearn.metrics import accuracy_score

model_looks_linear_HT_first = joblib.load('params_pkl/model_looks_linear_HT_first.pkl')
model_looks_rbf_HT_first = joblib.load('params_pkl/model_looks_rbf_ht_first.pkl')
model_looks_poly_HT_first = joblib.load('params_pkl/model_looks_poly_HT_first.pkl')

model_looks_linear_HT_second = joblib.load('params_pkl/model_looks_linear_HT_second.pkl')
model_looks_rbf_HT_second = joblib.load('params_pkl/model_looks_rbf_ht_second.pkl')
model_looks_poly_HT_second = joblib.load('params_pkl/model_looks_poly_HT_second.pkl')

#input the selected feature by using variance threshold
VT_selected_feature_Zoo = joblib.load('VT_selected_feature_Zoo.pkl')
pso_linear_first = joblib.load('PSO-selected-feature/pso-linear-first.pkl')
pso_rbf_first = joblib.load('PSO-selected-feature/pso-rbf-first.pkl')
pso_poly_first = joblib.load('PSO-selected-feature/pso-poly-first.pkl')

pso_linear_second = joblib.load('PSO-selected-feature/pso-linear-second.pkl')
pso_rbf_second = joblib.load('PSO-selected-feature/pso-rbf-second.pkl')
pso_poly_second = joblib.load('PSO-selected-feature/pso-poly-second.pkl')

#Input the dataset
X_train = pd.read_csv('dataset-after-preparation\X_train.csv')
X_test = pd.read_csv('dataset-after-preparation\X_test.csv')
y_train = pd.read_csv('dataset-after-preparation\y_train.csv')
y_test = pd.read_csv('dataset-after-preparation\y_test.csv')
X_valid = pd.read_csv('dataset-after-preparation\X_valid.csv')
y_valid = pd.read_csv('dataset-after-preparation\y_valid.csv')

# fill the dataset with the selected features
X_train = X_train[VT_selected_feature_Zoo]
X_test = X_test[VT_selected_feature_Zoo]

X_train.shape, X_test.shape,

# Feature Slicing with PSO-SVM
X_train_linear_first = X_train[pso_linear_first]
X_train_rbf_first = X_train[pso_rbf_first]
X_train_poly_first = X_train[pso_poly_first]

X_test_linear_first = X_test[pso_linear_first]
X_test_rbf_first = X_test[pso_rbf_first]
X_test_poly_first = X_test[pso_poly_first]

# Feature Slicing with PSO-SVM
X_train_linear_second = X_train[pso_linear_second]
X_train_rbf_second = X_train[pso_rbf_second]
X_train_poly_second = X_train[pso_poly_second]

X_test_linear_second = X_test[pso_linear_second]
X_test_rbf_second = X_test[pso_rbf_second]
X_test_poly_second = X_test[pso_poly_second]

# See the shape of Train data (First Feature)
X_train_linear_first.shape, X_train_rbf_first.shape, X_train_poly_first.shape

X_train_linear_second.shape, X_train_rbf_second.shape, X_train_poly_second.shape, 

### Making the prediction from the each kernel with them best estimator

# Test Data

## LNR1 

y_pred_test_linear_first = model_looks_linear_HT_first.predict(X_test_linear_first)
### score
print(f'LNR1 model test score: {accuracy_score(y_test, y_pred_test_linear_first)}')

### Test Dataset (For Confusion Matrix and Classification Report)

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_linear_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_linear_test = confusion_matrix(y_test, y_pred_test_linear_first, labels=[1, 0])
TN_linear_test = CM_linear_test[0][0]
FN_linear_test = CM_linear_test[1][0]
TP_linear_test = CM_linear_test[1][1]
FP_linear_test = CM_linear_test[0][1]
print("TN_linear_test: ", TN_linear_test)
print("FN_linear_test: ", FN_linear_test)
print("TP_linear_test: ", TP_linear_test)
print("FP_linear_test: ", FP_linear_test)

## LNR2

y_pred_test_linear_second = model_looks_linear_HT_second.predict(X_test_linear_second)
### score
print(f'LNR2 model test score: {accuracy_score(y_test, y_pred_test_linear_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_linear_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_linear_test = confusion_matrix(y_test, y_pred_test_linear_second, labels=[1, 0])
TN_linear_test = CM_linear_test[0][0]
FN_linear_test = CM_linear_test[1][0]
TP_linear_test = CM_linear_test[1][1]
FP_linear_test = CM_linear_test[0][1]
print("TN_linear_test: ", TN_linear_test)
print("FN_linear_test: ", FN_linear_test)
print("TP_linear_test: ", TP_linear_test)
print("FP_linear_test: ", FP_linear_test)

# RBF1

y_pred_test_rbf_first = model_looks_rbf_HT_first.predict(X_test_rbf_first)
### score
print(f'RBF1 model test score: {accuracy_score(y_test, y_pred_test_rbf_first)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_rbf_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_rbf_test = confusion_matrix(y_test, y_pred_test_rbf_first, labels=[1, 0])
TN_rbf_test = CM_rbf_test[0][0]
FN_rbf_test = CM_rbf_test[1][0]
TP_rbf_test = CM_rbf_test[1][1]
FP_rbf_test = CM_rbf_test[0][1]
print("TN_rbf_test: ", TN_rbf_test)
print("FN_rbf_test: ", FN_rbf_test)
print("TP_rbf_test: ", TP_rbf_test)
print("FP_rbf_test: ", FP_rbf_test)

# RBF2

y_pred_test_rbf_second = model_looks_rbf_HT_second.predict(X_test_rbf_second)
### score
print(f'RBF2 model test score: {accuracy_score(y_test, y_pred_test_rbf_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_rbf_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_rbf_test = confusion_matrix(y_test, y_pred_test_rbf_second, labels=[1, 0])
TN_rbf_test = CM_rbf_test[0][0]
FN_rbf_test = CM_rbf_test[1][0]
TP_rbf_test = CM_rbf_test[1][1]
FP_rbf_test = CM_rbf_test[0][1]
print("TN_rbf_test: ", TN_rbf_test)
print("FN_rbf_test: ", FN_rbf_test)
print("TP_rbf_test: ", TP_rbf_test)
print("FP_rbf_test: ", FP_rbf_test)

# POLY1

y_pred_test_poly_first = model_looks_poly_HT_first.predict(X_test_poly_first)
### score
print(f'POLY1 model test score: {accuracy_score(y_test, y_pred_test_poly_first)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_poly_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_poly_test = confusion_matrix(y_test, y_pred_test_poly_first, labels=[1, 0])
TN_poly_test = CM_poly_test[0][0]
FN_poly_test = CM_poly_test[1][0]
TP_poly_test = CM_poly_test[1][1]
FP_poly_test = CM_poly_test[0][1]
print("TN_poly_test: ", TN_poly_test)
print("FN_poly_test: ", FN_poly_test)
print("TP_poly_test: ", TP_poly_test)
print("FP_poly_test: ", FP_poly_test)

# POLY2

y_pred_test_poly_second = model_looks_poly_HT_second.predict(X_test_poly_second)
### score
print(f'POLY2 model test score: {accuracy_score(y_test, y_pred_test_poly_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_poly_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_poly_test = confusion_matrix(y_test, y_pred_test_poly_second, labels=[1, 0])
TN_poly_test = CM_poly_test[0][0]
FN_poly_test = CM_poly_test[1][0]
TP_poly_test = CM_poly_test[1][1]
FP_poly_test = CM_poly_test[0][1]
print("TN_poly_test: ", TN_poly_test)
print("FN_poly_test: ", FN_poly_test)
print("TP_poly_test: ", TP_poly_test)
print("FP_poly_test: ", FP_poly_test)

# Train data

y_pred_train_linear_first = model_looks_linear_HT_first.predict(X_train_linear_first)
### score
print(f'LNR1 model train score: {accuracy_score(y_train, y_pred_train_linear_first)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_train, y_pred_train_linear_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_linear_train = confusion_matrix(y_train, y_pred_train_linear_first, labels=[1, 0])
TN_linear_train = CM_linear_train[0][0]
FN_linear_train = CM_linear_train[1][0]
TP_linear_train = CM_linear_train[1][1]
FP_linear_train = CM_linear_train[0][1]
print("TN_linear_test: ", TN_linear_train)
print("FN_linear_test: ", FN_linear_train)
print("TP_linear_test: ", TP_linear_train)
print("FP_linear_test: ", FP_linear_train)

# LNR2

y_pred_train_linear_second = model_looks_linear_HT_second.predict(X_train_linear_second)
### score
print(f'LNR2 model train score: {accuracy_score(y_train, y_pred_train_linear_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred_test_linear_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_linear_train = confusion_matrix(y_train, y_pred_train_linear_second, labels=[1, 0])
TN_linear_train = CM_linear_train[0][0]
FN_linear_train = CM_linear_train[1][0]
TP_linear_train = CM_linear_train[1][1]
FP_linear_train = CM_linear_train[0][1]
print("TN_linear_test: ", TN_linear_train)
print("FN_linear_test: ", FN_linear_train)
print("TP_linear_test: ", TP_linear_train)
print("FP_linear_test: ", FP_linear_train)

# RBF1

y_pred_train_rbf_first = model_looks_rbf_HT_first.predict(X_train_rbf_first)
### score
print(f'RBF1 model train score: {accuracy_score(y_train, y_pred_train_rbf_first)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_train, y_pred_train_rbf_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_rbf_train = confusion_matrix(y_train, y_pred_train_rbf_first, labels=[1, 0])
TN_rbf_train = CM_rbf_train[0][0]
FN_rbf_train = CM_rbf_train[1][0]
TP_rbf_train = CM_rbf_train[1][1]
FP_rbf_train = CM_rbf_train[0][1]
print("TN_train_test: ", TN_rbf_train)
print("FN_train_test: ", FN_rbf_train)
print("TP_train_test: ", TP_rbf_train)
print("FP_train_test: ", FP_rbf_train)

# RBF2

y_pred_train_rbf_second = model_looks_rbf_HT_second.predict(X_train_rbf_second)
### score
print(f'RBF2 model train score: {accuracy_score(y_train, y_pred_train_rbf_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_train, y_pred_train_rbf_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_rbf_train = confusion_matrix(y_train, y_pred_train_rbf_second, labels=[1, 0])
TN_rbf_train = CM_rbf_train[0][0]
FN_rbf_train = CM_rbf_train[1][0]
TP_rbf_train = CM_rbf_train[1][1]
FP_rbf_train = CM_rbf_train[0][1]
print("TN_train_test: ", TN_rbf_train)
print("FN_train_test: ", FN_rbf_train)
print("TP_train_test: ", TP_rbf_train)
print("FP_train_test: ", FP_rbf_train)

# POLY1

y_pred_train_poly_first = model_looks_poly_HT_first.predict(X_train_poly_first)
### score
print(f'POLY1 model train score: {accuracy_score(y_train, y_pred_train_poly_first)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_train, y_pred_train_poly_first, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_poly_train = confusion_matrix(y_train, y_pred_train_poly_first, labels=[1, 0])
TN_poly_train = CM_poly_train[0][0]
FN_poly_train = CM_poly_train[1][0]
TP_poly_train = CM_poly_train[1][1]
FP_poly_train = CM_poly_train[0][1]
print("TN_train_test: ", TN_poly_train)
print("FN_train_test: ", FN_poly_train)
print("TP_train_test: ", TP_poly_train)
print("FP_train_test: ", FP_poly_train)

# POLY2

y_pred_train_poly_second = model_looks_poly_HT_second.predict(X_train_poly_second)
### score
print(f'POLY2 model train score: {accuracy_score(y_train, y_pred_train_poly_second)}')

# Classification Report for Linear kernel (Ttest set) 
from sklearn.metrics import classification_report
print(classification_report(y_train, y_pred_train_poly_second, labels=[1]))

# Confusion Matrix
from sklearn.metrics import confusion_matrix
CM_poly_train = confusion_matrix(y_train, y_pred_train_poly_second, labels=[1, 0])
TN_poly_train = CM_poly_train[0][0]
FN_poly_train = CM_poly_train[1][0]
TP_poly_train = CM_poly_train[1][1]
FP_poly_train = CM_poly_train[0][1]
print("TN_train_test: ", TN_poly_train)
print("FN_train_test: ", FN_poly_train)
print("TP_train_test: ", TP_poly_train)
print("FP_train_test: ", FP_poly_train)
